refactor: migrate profile and browse profile
